+++ tests/Norway/ConstitutionDayTest.php
@@ -67,7 +67,7 @@ class ConstitutionDayTest extends NorwayBaseTestCase implements YasumiTestCaseIn
             self::REGION,
             self::HOLIDAY,
             $this->generateRandomYear(self::ESTABLISHMENT_YEAR),
-            [self::LOCALE => 'Nasjonaldagen']
+            [self::LOCALE => 'Grunnlovsdagen']
         );
     }